pokeapi/requirements.txt
aiofiles==0.5.0
aiosqlite==0.16.1
aniso8601==7.0.0
appdirs==1.4.4
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.23.0
bcrypt==3.2.0
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.66.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
idna==3.2
iso8601==0.1.16
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
orjson==3.6.0
passlib==1.7.4
pathspec==0.8.1
promise==2.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pypika-tortoise==0.1.1
python-dotenv==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
starlette==0.14.2
tomli==1.0.4
tortoise-orm==0.17.5
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.15.3
watchgod==0.7
websockets==8.1
